whiskey_config/pyproject.toml
[project]
name = "whiskey-config"
version = "0.1.0"
description = "Configuration management extension for Whiskey framework"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "whiskey",
]

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]
toml = ["tomli>=2.0;python_version<'3.11'", "tomli-w>=1.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whiskey_config"]